+++ changelog.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 
+## [0.0.17] - 2023-09-26
+
+### Added
+
+- improved board class
+- changed board tests
+
 ## [0.0.16] - 2023-09-25
 
 ### Added